work towards using dtls 1.2
this is just trying the simplest possible set of changes (not using the
newer, non-deprecated APIs yet).  note that there is a bug in
bouncycastle pre 1.61 where an unknownhashalgorithm is thrown.  as of
this time, 1.61 hasn't made it into maven yet so this branch fails to
compile.  once 1.61 is in, try this code again to see if it works.  if
it doesn't, made need to move to the new bc APIs