change requestKeyframe API to not take in an SSRC.
this feels a bit weird, but my thinking is as follows:

we need to be able to request a keyframe without knowing a specific SSRC
for the purpose of requesting pre-emptive keyframes when the active
speaker switches.

this isn't a big deal, as currently there isn't a need to denote a
specific SSRC when asking for a keyframe: as long as we request it from
any of the video SSRCs we'll get keyframes for all the streams.  because
of this, requiring a specific SSRC be passed doesn't make a lot of
sense; so although i do still allow a specific one to be used in the
case of forwarding FIR/PLI (because it should make no difference and was
less work), in general requestKeyframe now doesn't take in an SSRC.

if/when in the future any of this changes, we'll be in the opposite
position: an SSRC will ALWAYS have to be passed in, so at that time we
can shift from this API back to one that requires an SSRC be passed--and
at that time we'll have to re-think how requesting pre-emptive keyframes
on active speaker switch should work.