fix concurrency issues in streaminformationstore.
now that nodes are accessing the fields held there directly (instead of
storing the data on their own) we need to make sure we don't run into
issues if they are updated while iterating.  note that we still need the
locks in streaminformationstore to ensure the container updates and
handler invocations happen atomically.