when creating a logger, explicitly use Class::class instead of this::class
this::class will work in most places, but it does not work correctly
when it's used for calling a parent class' constructor so i've changed
the usages to Class::class to be consistent.  Node's case is an
exception: we don't know the actual subtype, so we use this::class to
dynamically get the proper classname.  since we do this in the init
block it works fine.