fix: Expire the first incoming estimate after a period of inactivity.
We have observed cases where 1-1 calls go through the JVB briefly and
then they switch to p2p mode (as they normally should). The few packets
that go through the bridge are enough to produce a low first incoming
estimate.

Now, when a 3rd participant joins the call (long after the start of the
call) and media is routed through the bridge (as they normally should)
the aimd picks up where it left of with that stale first estimate. The
waiting time to produce a valid estimate (kInitializationTimeMs =
5000ms) is long past and so the aimd promotes that low first incoming
estimate as a valid estimate.

That drives the bitrate controller to suspend videos.

This PR implements a timeout for the first incoming estimate: If the
first incoming estimate is not promoted to a valid estimate in the first
kInitializationTimeMs*2, then we forget it and start over.